pyproject.toml
[project]
name = "rag-intro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.6",
    "openai>=1.97.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]
